+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added layers:
   - "Biology/Birds/Seabird colony regulated areas" from NatureMap
   - "Biology/Mammals/Polar bear denning areas" from NatureMap
+  - "Biology/Mammals/Musk oxen calving areas" from NatureMap
 
 # v4.0.0alpha1 (2026-03-04)